refactor(exec): move implementation of output functions from `run*` to `output*`
