+++ packages/exec/lib/index.d.ts
@@ -1,3 +1,4 @@
 export { Command } from "./command";
+export { output, outputSilently } from "./output";
 export { OutputResult, RunResult } from "./result";
-export { output, outputSilently, run, runSilently } from "./run";
+export { run, runSilently } from "./run";
