+++ packages/exec/lib/output.js
@@ -23,34 +23,34 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.runSilently = exports.run = void 0;
+exports.outputSilently = exports.output = void 0;
 const exec = __importStar(require("@actions/exec"));
 const result_1 = require("./result");
-async function runHelper(silent, command, ...args) {
-    const rc = await exec.exec(command, args, {
+async function outputHelper(silent, command, ...args) {
+    const res = await exec.getExecOutput(command, args, {
         ignoreReturnCode: true,
         silent,
     });
-    return new result_1.RunResult(rc);
+    return new result_1.OutputResult(res.exitCode, res.stdout);
 }
 /**
- * Runs a command
+ * Runs a command and gets the output
  * @param command a command to run
  * @param args additional arguments for the command
  * @returns a command run result
  */
-async function run(command, ...args) {
-    return runHelper(false, command, ...args);
+async function output(command, ...args) {
+    return outputHelper(false, command, ...args);
 }
-exports.run = run;
+exports.output = output;
 /**
- * Runs a command silently
+ * Runs a command silently and gets the output
  * @param command a command to run
  * @param args additional arguments for the command
  * @returns a command run result
  */
-async function runSilently(command, ...args) {
-    return runHelper(true, command, ...args);
+async function outputSilently(command, ...args) {
+    return outputHelper(true, command, ...args);
 }
-exports.runSilently = runSilently;
-//# sourceMappingURL=run.js.map
+exports.outputSilently = outputSilently;
+//# sourceMappingURL=output.js.map
