+++ packages/exec/lib/run.d.ts
@@ -1,4 +1,4 @@
-import { OutputResult, RunResult } from "./result";
+import { RunResult } from "./result";
 /**
  * Runs a command
  * @param command a command to run
@@ -13,17 +13,3 @@ export declare function run(command: string, ...args: string[]): Promise<RunResu
  * @returns a command run result
  */
 export declare function runSilently(command: string, ...args: string[]): Promise<RunResult>;
-/**
- * Runs a command and gets the output
- * @param command a command to run
- * @param args additional arguments for the command
- * @returns a command run result
- */
-export declare function output(command: string, ...args: string[]): Promise<OutputResult>;
-/**
- * Runs a command silently and gets the output
- * @param command a command to run
- * @param args additional arguments for the command
- * @returns a command run result
- */
-export declare function outputSilently(command: string, ...args: string[]): Promise<OutputResult>;
