+++ packages/exec/src/run.test.ts
@@ -1,5 +1,5 @@
 import { testRunOnSuccessAndFailed } from "./run-helper.test";
-import { output, outputSilently, run, runSilently } from "./run";
+import { run, runSilently } from "./run";
 
 testRunOnSuccessAndFailed({
   title: "runs a command",
@@ -25,35 +25,5 @@ testRunOnSuccessAndFailed({
     run: () => runSilently("node", "-e", "throw new Error('some error')"),
     runScript:
       "exec.runSilently('node', '-e', 'throw new Error(\"some error\")');",
-  },
-});
-
-testRunOnSuccessAndFailed({
-  title: "runs a command and gets the output",
-  shouldBeSilent: false,
-  onSuccess: {
-    run: () => output("node", "-e", "console.log('some log')"),
-    expectedOutput: "some log\n",
-    runScript: "exec.output('node', '-e', 'console.log(\"some log\")');",
-  },
-  onFailed: {
-    run: () => output("node", "-e", "throw new Error('some error')"),
-    runScript: "exec.output('node', '-e', 'throw new Error(\"some error\")');",
-  },
-});
-
-testRunOnSuccessAndFailed({
-  title: "runs a command silently and gets the output",
-  shouldBeSilent: true,
-  onSuccess: {
-    run: () => outputSilently("node", "-e", "console.log('some log')"),
-    expectedOutput: "some log\n",
-    runScript:
-      "exec.outputSilently('node', '-e', 'console.log(\"some log\")');",
-  },
-  onFailed: {
-    run: () => outputSilently("node", "-e", "throw new Error('some error')"),
-    runScript:
-      "exec.outputSilently('node', '-e', 'throw new Error(\"some error\")');",
   },
 });
