+++ packages/exec/src/run.ts
@@ -1,5 +1,5 @@
 import * as exec from "@actions/exec";
-import { OutputResult, RunResult } from "./result";
+import { RunResult } from "./result";
 
 async function runHelper(
   silent: boolean,
@@ -37,42 +37,4 @@ export async function runSilently(
   ...args: string[]
 ): Promise<RunResult> {
   return runHelper(true, command, ...args);
-}
-
-async function outputHelper(
-  silent: boolean,
-  command: string,
-  ...args: string[]
-): Promise<OutputResult> {
-  const res = await exec.getExecOutput(command, args, {
-    ignoreReturnCode: true,
-    silent,
-  });
-  return new OutputResult(res.exitCode, res.stdout);
-}
-
-/**
- * Runs a command and gets the output
- * @param command a command to run
- * @param args additional arguments for the command
- * @returns a command run result
- */
-export async function output(
-  command: string,
-  ...args: string[]
-): Promise<OutputResult> {
-  return outputHelper(false, command, ...args);
-}
-
-/**
- * Runs a command silently and gets the output
- * @param command a command to run
- * @param args additional arguments for the command
- * @returns a command run result
- */
-export async function outputSilently(
-  command: string,
-  ...args: string[]
-): Promise<OutputResult> {
-  return outputHelper(true, command, ...args);
 }
